refactor CLI as a module
